Draw screen banner characters, and remove pointer value indirection

+++ nome/src/settings.h
@@ -18,7 +18,7 @@ public:
     const size_t frames_per_period;
     
     uint16_t current_bpm;
-    uint16_t current_downbeat;
+    uint8_t current_downbeat;
 };
 
 }
